# kratos_client/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/create/estimate:
        post:
            tags:
                - Estimate
            operationId: Estimate_CreateEstimate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/.CreateEstimateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.CreateEstimateReply'
    /v1/delete/estimate:
        post:
            tags:
                - Estimate
            operationId: Estimate_DeleteEstimate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/.DeleteEstimateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.DeleteEstimateReply'
...
    /v1/get/estimate:
        post:
            tags:
                - Estimate
            operationId: Estimate_GetEstimate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/.GetEstimateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.GetEstimateReply'
    /v1/list/estimate:
        post:
            tags:
                - Estimate
            operationId: Estimate_ListDrugEstimate
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/.ListDrugEstimateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/.ListDrugEstimateReply'
    /v1/login:
        post:
            tags:
                - User
            operationId: User_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.LoginReply'
    /v1/sendSms:
        post:
            tags:
                - User
            description: Sends a greeting
            operationId: User_SendSms
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/user.v1.SendSmsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/user.v1.SendSmsReply'
components:
    schemas:
...
        .CreateEstimateReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
        .CreateEstimateRequest:
            type: object
            properties:
                drugId:
                    type: string
                userId:
                    type: string
                content:
                    type: string
                image:
                    type: string
        .DeleteEstimateReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
        .DeleteEstimateRequest:
            type: object
            properties:
                id:
                    type: string
        .EstimateInfo:
            type: object
            properties:
                drugId:
                    type: string
                content:
                    type: string
                image:
                    type: string
                mobile:
                    type: string
                createTime:
                    type: string
                avatar:
                    type: string
...
        .GetEstimateReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                info:
                    $ref: '#/components/schemas/.EstimateInfo'
        .GetEstimateRequest:
            type: object
            properties:
                id:
                    type: string
                userId:
                    type: string
...
        .InfoDrugs:
            type: object
            properties:
                drugName:
                    type: string
                specification:
                    type: string
                price:
                    type: number
                    format: double
                salesVolume:
                    type: number
                    format: double
                inventory:
                    type: string
                exhibitionUrl:
                    type: string
        .ListDrugEstimateReply:
            type: object
            properties:
                code:
                    type: string
                message:
                    type: string
                info:
                    $ref: '#/components/schemas/.InfoDrugs'
                list:
                    type: array
                    items:
                        $ref: '#/components/schemas/.EstimateInfo'
        .ListDrugEstimateRequest:
            type: object
            properties:
                drugId:
                    type: string
                userId:
                    type: string
...
        user.v1.LoginReply:
            type: object
            properties:
                message:
                    type: string
                token:
                    type: string
            description: The response message containing the greetings
        user.v1.LoginRequest:
            type: object
            properties:
                mobile:
                    type: string
                sendSmsCode:
                    type: string
            description: The request message containing the user's name.
        user.v1.SendSmsReply:
            type: object
            properties:
                message:
                    type: string
            description: The response message containing the greetings
        user.v1.SendSmsRequest:
            type: object
            properties:
                mobile:
                    type: string
                source:
                    type: string
            description: The request message containing the user's name.
...